Moved tiles to 'blocks' folder.
